Move the `codec` module under the `@bayou` namespace.
