+++ local-modules/@bayou/codec/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "codec",
+  "name": "@bayou/codec",
   "version": "1.0.0",
 
   "dependencies": {
