+++ local-modules/@bayou/codec/tests/test_Codec_encode.js
@@ -5,8 +5,8 @@
 import { assert } from 'chai';
 import { describe, it } from 'mocha';
 
-import { Codec, ConstructorCall } from 'codec';
-import { MockCodable } from 'codec/mocks';
+import { Codec, ConstructorCall } from '@bayou/codec';
+import { MockCodable } from '@bayou/codec/mocks';
 import { FrozenBuffer } from '@bayou/util-common';
 
 describe('api-common/Codec.encode*()', () => {
