+++ local-modules/api-server/package.json
@@ -6,7 +6,7 @@
     "content-type": "^1.0.2",
 
     "api-common": "local",
-    "codec": "local",
+    "@bayou/codec": "local",
     "hooks-server": "local",
     "promise-util": "local",
     "@bayou/see-all": "local",
