+++ local-modules/app-common/package.json
@@ -4,7 +4,7 @@
 
   "dependencies": {
     "api-common": "local",
-    "codec": "local",
+    "@bayou/codec": "local",
     "doc-common": "local",
     "ot-common": "local",
     "@bayou/util-common": "local"
