+++ local-modules/doc-server/package.json
@@ -6,7 +6,7 @@
     "weak": "^1.0.1",
 
     "app-common": "local",
-    "codec": "local",
+    "@bayou/codec": "local",
     "doc-common": "local",
     "env-server": "local",
     "file-store": "local",
