+++ local-modules/file-store/FileCodec.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { Codec } from 'codec';
+import { Codec } from '@bayou/codec';
 import { TransactionOp } from 'file-store-ot';
 import { CommonBase } from '@bayou/util-common';
