+++ local-modules/file-store/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "weak": "^1.0.1",
 
-    "codec": "local",
+    "@bayou/codec": "local",
     "file-store-ot": "local",
     "hooks-server": "local",
     "promise-util": "local",
